@@ -1,31 +1,41 @@
-'use strict';
+"use strict";
 
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
+exports["default"] = void 0;
 
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
+var _react = _interopRequireWildcard(require("react"));
 
-var _react = require('react');
+var _propTypes = _interopRequireDefault(require("prop-types"));
 
-var _react2 = _interopRequireDefault(_react);
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
-var _propTypes = require('prop-types');
+function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {}; if (desc.get || desc.set) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } } newObj["default"] = obj; return newObj; } }
 
-var _propTypes2 = _interopRequireDefault(_propTypes);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+function _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === "object" || typeof call === "function")) { return call; } return _assertThisInitialized(self); }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+function _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return self; }
+
+function _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }
+
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
+
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }
+
+function _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }
+
+function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
 
 var nil = function nil() {
   return false;
 };
-
 /*
  * Parabola 组件参数说明：
  * - start 起始点坐标/宽高
@@ -39,37 +49,36 @@ var nil = function nil() {
  * - onAfterEnd 动画销毁之后事件
  */
 
-var Parabola = function (_Component) {
+
+var Parabola =
+/*#__PURE__*/
+function (_Component) {
   _inherits(Parabola, _Component);
 
   _createClass(Parabola, null, [{
-    key: 'calculate',
-
-
-    /*
-     * 根据极值点和另外一点计算抛物线参数
-     * 抛物线顶点式方程：y=a(x-h)^2+k，其中 (h, k) 为 极值点坐标
-     */
+    key: "calculate",
     value: function calculate(extreme, point) {
       var ex = extreme.x,
           ey = extreme.y;
       var x = point.x,
           y = point.y;
-
-
       var a = (y - ey) / Math.pow(x - ex, 2);
       var h = ex;
       var k = ey;
-
-      return { a: a, h: h, k: k };
+      return {
+        a: a,
+        h: h,
+        k: k
+      };
     }
   }]);
 
   function Parabola(props) {
+    var _this;
+
     _classCallCheck(this, Parabola);
 
-    // 初始状态
-    var _this = _possibleConstructorReturn(this, (Parabola.__proto__ || Object.getPrototypeOf(Parabola)).call(this, props));
+    _this = _possibleConstructorReturn(this, _getPrototypeOf(Parabola).call(this, props)); // 初始状态
 
     _this.state = {
       translateX: 0,
@@ -78,98 +87,109 @@ var Parabola = function (_Component) {
       scaleY: 1,
       animationEnd: false
     };
-
     _this.startTime = new Date();
 
     _this.initParabolas();
+
     return _this;
   }
 
   _createClass(Parabola, [{
-    key: 'componentDidMount',
+    key: "componentDidMount",
     value: function componentDidMount() {
       this.beforeStart();
       this.draw();
     }
   }, {
-    key: 'componentWillUnmount',
+    key: "componentWillUnmount",
     value: function componentWillUnmount() {
       this.afterEnd();
     }
   }, {
-    key: 'draw',
+    key: "draw",
     value: function draw() {
       var _this2 = this;
 
+      var delay = this.props.delay;
       setTimeout(function () {
         _this2.update();
-        _this2.startTime = new Date();
-      }, this.props.delay);
-    }
-  }, {
-    key: 'finish',
-    value: function finish() {
-      if (this.props.onEnd) {
-        this.props.onEnd();
-      }
-    }
-  }, {
-    key: 'beforeStart',
-    value: function beforeStart() {
-      if (this.props.onBeforeStart) {
-        this.props.onBeforeStart();
-      }
-    }
-  }, {
-    key: 'afterEnd',
-    value: function afterEnd() {
-      if (this.props.onAfterEnd) {
-        this.props.onAfterEnd();
-      }
-    }
 
+        _this2.startTime = new Date();
+      }, delay);
+    }
+  }, {
+    key: "finish",
+    value: function finish() {
+      var onEnd = this.props.onEnd;
+
+      if (onEnd) {
+        onEnd();
+      }
+    }
+  }, {
+    key: "beforeStart",
+    value: function beforeStart() {
+      var onBeforeStart = this.props.onBeforeStart;
+
+      if (onBeforeStart) {
+        onBeforeStart();
+      }
+    }
+  }, {
+    key: "afterEnd",
+    value: function afterEnd() {
+      var onAfterEnd = this.props.onAfterEnd;
+
+      if (onAfterEnd) {
+        onAfterEnd();
+      }
+    }
     /*
      * 坐标系统转换，将全局坐标系转换为相对坐标系
      */
 
   }, {
-    key: 'initParabolas',
+    key: "initParabolas",
     value: function initParabolas() {
-      var _props = this.props,
-          rate = _props.rate,
-          top = _props.top,
-          start = _props.start,
-          end = _props.end;
-
-      var left = { // 起点
+      var _this$props = this.props,
+          rate = _this$props.rate,
+          top = _this$props.top,
+          start = _this$props.start,
+          end = _this$props.end;
+      var left = {
+        // 起点
         x: 0,
         y: 0
       };
-      var right = { // 终点
+      var right = {
+        // 终点
         x: end.x - start.x,
         y: end.y - start.y
       };
-      var extreme = { // 极值点
+      var extreme = {
+        // 极值点
         x: rate * right.x,
         y: Math.min(right.y, left.y) - top
       };
-
       var lParabola = Parabola.calculate(extreme, left); // 极值点左侧抛物线参数
+
       var rParabola = Parabola.calculate(extreme, right); // 极值点右侧抛物线参数
 
-      this.parabolas = { lParabola: lParabola, rParabola: rParabola };
+      this.parabolas = {
+        lParabola: lParabola,
+        rParabola: rParabola
+      };
       return this.parabolas;
     }
   }, {
-    key: 'update',
+    key: "update",
     value: function update() {
-      var _props2 = this.props,
-          duration = _props2.duration,
-          rate = _props2.rate,
-          start = _props2.start,
-          end = _props2.end;
+      var _this$props2 = this.props,
+          duration = _this$props2.duration,
+          rate = _this$props2.rate,
+          start = _this$props2.start,
+          end = _this$props2.end;
       var animationEnd = this.state.animationEnd;
-
       var interval = Date.now() - this.startTime;
 
       if (interval > duration) {
@@ -182,72 +202,63 @@ var Parabola = function (_Component) {
       }
 
       var percent = interval / duration;
-      var _parabolas = this.parabolas,
-          lParabola = _parabolas.lParabola,
-          rParabola = _parabolas.rParabola;
+      var _this$parabolas = this.parabolas,
+          lParabola = _this$parabolas.lParabola,
+          rParabola = _this$parabolas.rParabola;
 
       var _ref = rate > percent ? lParabola : rParabola,
           a = _ref.a,
           h = _ref.h,
-          k = _ref.k;
-
-      // 计算位置
+          k = _ref.k; // 计算位置
 
 
       var x = percent * (end.x - start.x);
-      var y = a * Math.pow(x - h, 2) + k;
+      var y = a * Math.pow(x - h, 2) + k; // 计算缩放
 
-      // 计算缩放
       var scaleX = 1 - percent * (1 - (end.width || 1) / (start.width || 1));
       var scaleY = 1 - percent * (1 - (end.height || 1) / (start.height || 1));
-
       this.setState({
         translateX: x,
         translateY: y,
         scaleX: scaleX,
         scaleY: scaleY,
         animationEnd: animationEnd
-      });
+      }); // 刷新
 
-      // 刷新
       requestAnimationFrame(this.update.bind(this));
-
       return false;
     }
   }, {
-    key: 'render',
+    key: "render",
     value: function render() {
-      var _state = this.state,
-          translateX = _state.translateX,
-          translateY = _state.translateY,
-          scaleX = _state.scaleX,
-          scaleY = _state.scaleY;
-      var _props3 = this.props,
-          start = _props3.start,
-          children = _props3.children;
-
-      return _react2.default.createElement(
-        'div',
-        {
-          style: {
-            position: 'absolute',
-            left: start.x + 'px',
-            top: start.y + 'px',
-            width: start.width + 'px',
-            height: start.height + 'px',
-            transform: 'translate(' + translateX + 'px, ' + translateY + 'px) scale(' + scaleX + ', ' + scaleY + ')',
-            transformOrigin: '0 0'
-          }
-        },
-        children
-      );
+      var _this$state = this.state,
+          translateX = _this$state.translateX,
+          translateY = _this$state.translateY,
+          scaleX = _this$state.scaleX,
+          scaleY = _this$state.scaleY;
+      var _this$props3 = this.props,
+          start = _this$props3.start,
+          children = _this$props3.children;
+      return _react["default"].createElement("div", {
+        style: {
+          position: 'absolute',
+          left: "".concat(start.x, "px"),
+          top: "".concat(start.y, "px"),
+          width: "".concat(start.width, "px"),
+          height: "".concat(start.height, "px"),
+          transform: "translate(".concat(translateX, "px, ").concat(translateY, "px) scale(").concat(scaleX, ", ").concat(scaleY, ")"),
+          transformOrigin: '0 0'
+        }
+      }, children);
     }
   }]);
 
   return Parabola;
 }(_react.Component);
 
-Parabola.defaultProps = {
+exports["default"] = Parabola;
+
+_defineProperty(Parabola, "defaultProps", {
   rate: 1,
   duration: 800,
   top: 0,
@@ -256,26 +267,32 @@ Parabola.defaultProps = {
   onEnd: nil,
   onBeforeStart: nil,
   onAfterEnd: nil
-};
-Parabola.propTypes = {
-  rate: _propTypes2.default.number,
-  start: _propTypes2.default.shape({
-    x: _propTypes2.default.number.isRequired,
-    y: _propTypes2.default.number.isRequired,
-    width: _propTypes2.default.number.isRequire,
-    height: _propTypes2.default.number.isRequire
+});
+
+_defineProperty(Parabola, "propTypes", {
+  rate: _propTypes["default"].number,
+  start: _propTypes["default"].shape({
+    x: _propTypes["default"].number.isRequired,
+    y: _propTypes["default"].number.isRequired,
+    width: _propTypes["default"].number.isRequired,
+    height: _propTypes["default"].number.isRequired
   }).isRequired,
-  end: _propTypes2.default.shape({
-    x: _propTypes2.default.number.isRequired,
-    y: _propTypes2.default.number.isRequired,
-    width: _propTypes2.default.number.isRequire,
-    height: _propTypes2.default.number.isRequire
+  end: _propTypes["default"].shape({
+    x: _propTypes["default"].number.isRequired,
+    y: _propTypes["default"].number.isRequired,
+    width: _propTypes["default"].number.isRequired,
+    height: _propTypes["default"].number.isRequired
   }).isRequired,
-  duration: _propTypes2.default.number,
-  top: _propTypes2.default.number,
-  delay: _propTypes2.default.number,
-  children: _propTypes2.default.element,
-  onEnd: _propTypes2.default.func,
-  onBeforeStart: _propTypes2.default.func,
-  onAfterEnd: _propTypes2.default.func };
-exports.default = Parabola;
+  duration: _propTypes["default"].number,
+  top: _propTypes["default"].number,
+  delay: _propTypes["default"].number,
+  children: _propTypes["default"].element,
+  onEnd: _propTypes["default"].func,
+  onBeforeStart: _propTypes["default"].func,
+  onAfterEnd: _propTypes["default"].func
+  /*
+   * 根据极值点和另外一点计算抛物线参数
+   * 抛物线顶点式方程：y=a(x-h)^2+k，其中 (h, k) 为 极值点坐标
+   */
+
+});
