@@ -1,31 +1,34 @@
-'use strict';
+"use strict";
 
-var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.showParabola = showParabola;
+exports["default"] = void 0;
 
-var _react = require('react');
+var _react = _interopRequireDefault(require("react"));
 
-var _react2 = _interopRequireDefault(_react);
+var _reactDom = _interopRequireDefault(require("react-dom"));
 
-var _reactDom = require('react-dom');
+var _index = _interopRequireDefault(require("./components/index"));
 
-var _reactDom2 = _interopRequireDefault(_reactDom);
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
-var _index = require('./components/index');
+function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; var ownKeys = Object.keys(source); if (typeof Object.getOwnPropertySymbols === 'function') { ownKeys = ownKeys.concat(Object.getOwnPropertySymbols(source).filter(function (sym) { return Object.getOwnPropertyDescriptor(source, sym).enumerable; })); } ownKeys.forEach(function (key) { _defineProperty(target, key, source[key]); }); } return target; }
 
-var _index2 = _interopRequireDefault(_index);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
 
 var nil = function nil() {};
 
 function showParabola(config) {
-  var props = _extends({}, config);
+  var props = _objectSpread({}, config);
 
   var div = document.createElement('div');
   document.body.appendChild(div);
 
   function remove() {
-    var unmountResult = _reactDom2.default.unmountComponentAtNode(div);
+    var unmountResult = _reactDom["default"].unmountComponentAtNode(div);
+
     if (unmountResult && div.parentNode) {
       div.parentNode.removeChild(div);
     }
@@ -39,36 +42,32 @@ function showParabola(config) {
       children = props.children,
       delay = props.delay,
       _props$onAfterEnd = props.onAfterEnd,
-      _onAfterEnd = _props$onAfterEnd === undefined ? nil : _props$onAfterEnd,
+      _onAfterEnd = _props$onAfterEnd === void 0 ? nil : _props$onAfterEnd,
       _props$onBeforeStart = props.onBeforeStart,
-      _onBeforeStart = _props$onBeforeStart === undefined ? nil : _props$onBeforeStart;
+      _onBeforeStart = _props$onBeforeStart === void 0 ? nil : _props$onBeforeStart;
 
-  _reactDom2.default.render(_react2.default.createElement(
-    _index2.default,
-    {
-      duration: duration,
-      rate: rate,
-      delay: delay,
-      top: top,
-      start: start,
-      end: end,
-      onBeforeStart: function onBeforeStart() {
-        return _onBeforeStart();
-      },
-      onEnd: function onEnd() {
-        return remove();
-      },
-      onAfterEnd: function onAfterEnd() {
-        return _onAfterEnd();
-      }
+  _reactDom["default"].render(_react["default"].createElement(_index["default"], {
+    duration: duration,
+    rate: rate,
+    delay: delay,
+    top: top,
+    start: start,
+    end: end,
+    onBeforeStart: function onBeforeStart() {
+      return _onBeforeStart();
     },
-    children
-  ), div);
+    onEnd: function onEnd() {
+      return remove();
+    },
+    onAfterEnd: function onAfterEnd() {
+      return _onAfterEnd();
+    }
+  }, children), div);
 }
 
-module.exports = {
+var _default = {
   showParabola: showParabola,
-  Parabola: _index2.default
-};
+  Parabola: _index["default"]
+}; // module.export { showParabola }, { Parabola };
 
-// module.export { showParabola }, { Parabola };
+exports["default"] = _default;
